Revert chars_format to an enum class

+++ include/boost/charconv/detail/fast_float/ascii_number.hpp
@@ -197,7 +197,7 @@ parsed_number_string_t<UC> parse_number_string(UC const *p, UC const * pend, par
     return answer;
   }
   int64_t exp_number = 0;            // explicit exponential part
-  if ((fmt & chars_format::scientific) && (p != pend) && ((UC('e') == *p) || (UC('E') == *p))) {
+  if (((unsigned)fmt & (unsigned)chars_format::scientific) && (p != pend) && ((UC('e') == *p) || (UC('E') == *p))) {
     UC const * location_of_e = p;
     ++p;
     bool neg_exp = false;
@@ -208,7 +208,7 @@ parsed_number_string_t<UC> parse_number_string(UC const *p, UC const * pend, par
       ++p;
     }
     if ((p == pend) || !is_integer(*p)) {
-      if(!(fmt & chars_format::fixed)) {
+      if(!((unsigned)fmt & (unsigned)chars_format::fixed)) {
         // We are in error.
         return answer;
       }
@@ -227,7 +227,10 @@ parsed_number_string_t<UC> parse_number_string(UC const *p, UC const * pend, par
     }
   } else {
     // If it scientific and not fixed, we have to bail out.
-    if((fmt & chars_format::scientific) && !(fmt & chars_format::fixed)) { return answer; }
+    if(((unsigned)fmt & (unsigned)chars_format::scientific) && !((unsigned)fmt & (unsigned)chars_format::fixed))
+    {
+        return answer;
+    }
   }
   answer.lastmatch = p;
   answer.valid = true;
